replace swagger configuration to extension method, add description of register method, change error handling middleware.
