+++ src/API/Startup.cs
@@ -41,28 +41,7 @@ namespace API
             services.AddAuthorizationService();
 
             services.AddControllers();
-            // Register the Swagger generator, defining 1 or more Swagger documents
-            services.AddSwaggerGen(c =>
-            {
-                c.SwaggerDoc("v1", new OpenApiInfo
-                {
-                    Version = "v1",
-                    Title = "CarRentalSystem API",
-                    Description = "Web api to rent cars",
-                    TermsOfService = new Uri("https://example.com/BestCarRentalSystem"),
-                    Contact = new OpenApiContact
-                    {
-                        Name = "Dmitry Karkanitsa",
-                        Email = "aakarkanica@gmail.com",
-                        Url = new Uri("https://vk.com/dkarkanitsa")
-                    }
-                });
-
-                // Set the comments path for the Swagger JSON and UI.
-                var xmlFile = $"{Assembly.GetExecutingAssembly().GetName().Name}.xml";
-                var xmlPath = Path.Combine(AppContext.BaseDirectory, xmlFile);
-                c.IncludeXmlComments(xmlPath);
-            });
+            services.AddSwaggerGenerator();
         }
 
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
@@ -77,9 +56,6 @@ namespace API
             }
 
             app.UseSwagger();
-
-            // Enable middleware to serve swagger-ui (HTML, JS, CSS, etc.),
-            // specifying the Swagger JSON endpoint.
             app.UseSwaggerUI(c =>
             {
                 c.SwaggerEndpoint("/swagger/v1/swagger.json", "My API V1");
